Test: cleaded up some code and created a dbUtils file to handle database emulators
