+++ src/__tests__/Quiz.test.jsx
@@ -1,5 +1,4 @@
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
-import React from "react";
 import Quiz from "../pages/quiz/Quiz";
 import { it, expect, describe, vi, beforeEach, beforeAll } from "vitest";
 import {
@@ -10,8 +9,7 @@ import {
   useLocation,
 } from "react-router-dom";
 import { useAuthContext } from "@/AuthProvider";
-import { db } from "./setup";
-import { doc, setDoc } from "firebase/firestore";
+import { setUpData, cleanUpData } from "./dbUtils";
 
 // Mock the useAuthContext hook
 vi.mock("@/AuthProvider", () => {
@@ -49,136 +47,6 @@ const renderComponent = (userRole, userEmail, TargetNum) => {
   );
 };
 
-const setUpData = async () => {
-  await setDoc(doc(db, "learners", "learner1"), {
-    email: "test_learner@example.com",
-    firstName: "Test",
-    lastname: "Learner",
-    scores: { sdg11t1: 80, sdg11t2: 100, sdg11t3: 0 },
-  });
-  await setDoc(doc(db, "learners", "learner2"), {
-    email: "learner_test@example.com",
-    firstName: "learner_two",
-    lastname: "New",
-    scores: { sdg11t1: 0, sdg11t2: 0, sdg11t3: 0 },
-  });
-  await setDoc(doc(db, "admins", "admin1"), {
-    email: "test_admin@example.com",
-    firstName: "Test",
-    lastname: "Admin",
-  });
-  const testQuizdata = [
-    {
-      quizId: "sdg11t1",
-      questions: [
-        {
-          id: "Q1",
-          questionText: "What is the capital of France?",
-          options: ["London", "Kenya", "Paris", "Dublin"],
-          correctAnswers: ["Paris"],
-          type: "mcq",
-        },
-        {
-          id: "Q2",
-          questionText: "What is the capital of Germany?",
-          options: ["Sydney", "Berlin", "Jakarta", "Chennai"],
-          correctAnswers: ["Berlin"],
-          type: "mcq",
-        },
-        {
-          id: "Q3",
-          questionText: "Why was 6 scared of 7?",
-          options: ["because", "7", "8", "9"],
-          correctAnswers: ["7", "8", "9"],
-          type: "ms",
-        },
-      ],
-    },
-    {
-      quizId: "sdg11t2",
-      questions: [
-        {
-          id: "Q1",
-          questionText: "What does the fox say?",
-          options: [
-            "Ring-ding-ding-ding-dingeringeding!",
-            "Wa-pa-pa-pa-pa-pa-pow!",
-            "Joff-tchoff-tchoffo-tchoffo-tchoff!",
-            "None of the above",
-          ],
-          correctAnswers: ["None of the above"],
-          type: "mcq",
-        },
-        {
-          id: "Q2",
-          questionText: "Who let the dogs out?",
-          options: [
-            "Baha Men",
-            "Snoop Dog",
-            "ur mum",
-            "the dog let itself out",
-          ],
-          correctAnswers: ["Baha Men"],
-          type: "mcq",
-        },
-        {
-          id: "Q3",
-          questionText: "This is a multiple select question",
-          options: ["pick", "multiple", "answers", "another answer"],
-          correctAnswers: ["pick", "multiple", "answers"],
-          type: "ms",
-        },
-      ],
-    },
-    {
-      quizId: "sdg11t3",
-      questions: [
-        {
-          id: "Q1",
-          questionText: "Hotel?",
-          options: ["Trivago", "Tripadvisor", "airbnb", "booking.com"],
-          correctAnswers: ["Trivago"],
-          type: "mcq",
-        },
-        {
-          id: "Q2",
-          questionText: "What is the capital of Germany?",
-          options: ["London", "Berlin", "Paris", "Dublin"],
-          correctAnswers: ["Berlin"],
-          type: "mcq",
-        },
-        {
-          id: "Q3",
-          questionText: "Can I be bothered thinking of another question?",
-          options: [
-            "clearly not",
-            "not correct",
-            "also not correct",
-            "I should have used the Qs we alaredy had",
-          ],
-          correctAnswers: [
-            "I should have used the Qs we alaredy had",
-            "clearly not",
-          ],
-          type: "ms",
-        },
-      ],
-    },
-  ];
-
-  for (const quiz of testQuizdata) {
-    const quizRef = doc(db, "quizzes", quiz.quizId);
-    for (const question of quiz.questions) {
-      await setDoc(doc(quizRef, "questions", question.id), {
-        questionText: question.questionText,
-        options: question.options,
-        correctAnswers: question.correctAnswers,
-        type: question.type,
-      });
-    }
-  }
-};
-
 describe("Quiz Component", () => {
   // Set up the database before all tests
   beforeAll(async () => {
@@ -188,6 +56,10 @@ describe("Quiz Component", () => {
   // Clear mocks before each test
   beforeEach(() => {
     vi.clearAllMocks();
+  });
+
+  afterAll(async () => {
+    await cleanUpData();
   });
 
   it("should render the score page if the learner has already has a score,", async () => {
